test: more coverage
