+++ tui/actions_test.go
@@ -492,54 +492,3 @@ func TestExitViewFile(t *testing.T) {
 
 	assert.False(t, ui.pages.HasPage("file"))
 }
-
-func TestExportAnalysis(t *testing.T) {
-	parentDir := &analyze.Dir{
-		File: &analyze.File{
-			Name: "parent",
-		},
-		Files: make([]fs.Item, 0, 1),
-	}
-	currentDir := &analyze.Dir{
-		File: &analyze.File{
-			Name:   "sub",
-			Parent: parentDir,
-		},
-	}
-
-	simScreen := testapp.CreateSimScreen()
-	defer simScreen.Fini()
-
-	app := testapp.CreateMockedApp(true)
-	ui := CreateUI(app, simScreen, &bytes.Buffer{}, false, true, false, false, false)
-	ui.done = make(chan struct{})
-	ui.Analyzer = &testanalyze.MockedAnalyzer{}
-	ui.currentDir = currentDir
-	ui.topDir = parentDir
-
-	assert.Equal(t, "sub", ui.currentDir.GetName())
-
-	ui.keyPressed(tcell.NewEventKey(tcell.KeyRune, 'E', 0))
-
-	assert.True(t, ui.pages.HasPage("export"))
-
-	ui.keyPressed(tcell.NewEventKey(tcell.KeyEnter, 0, 0))
-
-	assert.True(t, ui.pages.HasPage("export"))
-
-	// we cannot send Enter to the form from here
-
-	ui.exportAnalysis()
-
-	assert.True(t, ui.pages.HasPage("exporting"))
-
-	<-ui.done
-
-	assert.FileExists(t, "export.json")
-	err := os.Remove("export.json")
-	assert.NoError(t, err)
-
-	for _, f := range ui.app.(*testapp.MockedApp).GetUpdateDraws() {
-		f()
-	}
-}
